+++ v3/deprecated/20241205-v3-composite-liquidity-router/output/base.json
@@ -1,0 +1,3 @@
+{
+  "CompositeLiquidityRouter": "0xf23b4DB826DbA14c0e857029dfF076b1c0264843"
+}
