+++ v3/deprecated/20241205-v3-composite-liquidity-router/output/gnosis.json
@@ -1,0 +1,3 @@
+{
+  "CompositeLiquidityRouter": "0xC1A64500E035D9159C8826E982dFb802003227f0"
+}
