+++ v3/deprecated/20241205-v3-composite-liquidity-router/output/mainnet.json
@@ -1,0 +1,3 @@
+{
+  "CompositeLiquidityRouter": "0x1CD776897ef4f647bf8241Ec69549e4A9cb1D608"
+}
