+++ v3/deprecated/20241205-v3-composite-liquidity-router/output/sepolia.json
@@ -1,0 +1,3 @@
+{
+  "CompositeLiquidityRouter": "0xc6674C0c7694E9b990eAc939E74F8cc3DD39B4b0"
+}
