--- v3/tasks/20241205-v3-composite-liquidity-router/output/arbitrum.json
@@ -1,3 +1,0 @@
-{
-  "CompositeLiquidityRouter": "0x1311Fbc9F60359639174c1e7cC2032DbDb5Cc4d1"
-}
