--- v3/tasks/20241205-v3-composite-liquidity-router/output/sepolia.json
@@ -1,3 +1,0 @@
-{
-  "CompositeLiquidityRouter": "0xc6674C0c7694E9b990eAc939E74F8cc3DD39B4b0"
-}
