--- v3/tasks/20241205-v3-composite-liquidity-router/readme.md
@@ -1,15 +1,0 @@
-# 2024-12-05 - V3 Composite Liquidity Router
-
-Composite Liquidity Router deployment for V3.
-Contains `CompositeLiquidityRouter` for complex liquidity operations involving pools with ERC4626 wrappers and nested pools.
-
-## Useful Files
-
-- [Code](https://github.com/balancer/balancer-v3-monorepo/commit/74d7068fb21565741427cdabfa4f1b539a4bddaa).
-- [Ethereum mainnet addresses](./output/mainnet.json)
-- [Gnosis mainnet addresses](./output/gnosis.json)
-- [Arbitrum mainnet addresses](./output/arbitrum.json)
-- [Base mainnet addresses](./output/base.json)
-- [Sepolia testnet addresses](./output/sepolia.json)
-- [`CompositeLiquidityRouter` artifact](./artifact/CompositeLiquidityRouter.json)
-
